+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- `Innmind\MediaType\MediaType` constructor is now private, use `::from()` instead
+
 ## 2.2.0 - 2023-09-16
 
 ### Added